+++ phonopy/exception.py
@@ -45,3 +45,15 @@ class ForceCalculatorRequiredError(RuntimeError):
     """Exception when force calculator is required to compute force constants."""
 
     pass
+
+
+class CellNotFoundError(RuntimeError):
+    """Exception when unit cell not found."""
+
+    pass
+
+
+class MagmomValueError(ValueError):
+    """Exception when magnetic moment value is not valid."""
+
+    pass