+++ phonopy/interface/calculator.py
@@ -35,15 +35,16 @@
 # POSSIBILITY OF SUCH DAMAGE.
 from __future__ import annotations
 
+import os
 import pathlib
 import warnings
 from argparse import ArgumentParser
 from collections.abc import Sequence
 from math import pi, sqrt
-from typing import Optional, Union
 
 import numpy as np
 import yaml
+from numpy.typing import NDArray
 
 from phonopy.file_IO import get_supported_file_extensions_for_compression
 from phonopy.interface.phonopy_yaml import PhonopyYaml
@@ -64,6 +65,7 @@ calculator_info = {
     "elk": {"option": {"name": "--elk", "help": "Invoke elk mode"}},
     "fleur": {"option": {"name": "--fleur", "help": "Invoke Fleur mode"}},
     "lammps": {"option": {"name": "--lammps", "help": "Invoke Lammps mode"}},
+    "qlm": {"option": {"name": "--qlm", "help": "Invoke Questaal/LMTO mode"}},
     "qe": {"option": {"name": "--qe", "help": "Invoke Quantum espresso (QE) mode"}},
     "siesta": {"option": {"name": "--siesta", "help": "Invoke Siesta mode"}},
     "turbomole": {"option": {"name": "--turbomole", "help": "Invoke TURBOMOLE mode"}},
@@ -124,10 +126,10 @@ def convert_crystal_structure(
 
 
 def write_crystal_structure(
-    filename: Optional[str],
+    filename: str | os.PathLike | None,
     cell: PhonopyAtoms,
-    interface_mode: Optional[str] = None,
-    optional_structure_info: Optional[tuple] = None,
+    interface_mode: str | None = None,
+    optional_structure_info: tuple | None = None,
 ):
     """Write crystal structure to file in each calculator format.
 
@@ -251,6 +253,12 @@ def write_crystal_structure(
         import phonopy.interface.lammps as lammps
 
         lammps.write_lammps(filename, cell)
+
+    elif interface_mode == "qlm":
+        import phonopy.interface.qlm as qlm
+
+        qlm.write_qlm(filename, cell)
+
     elif interface_mode == "pwmat":
         import phonopy.interface.pwmat as pwmat
 
@@ -260,13 +268,13 @@ def write_crystal_structure(
 
 
 def write_supercells_with_displacements(
-    interface_mode: str,
+    interface_mode: str | None,
     supercell: PhonopyAtoms,
     cells_with_disps: Sequence[PhonopyAtoms],
-    optional_structure_info: Optional[tuple] = None,
-    displacement_ids: Optional[Union[Sequence, np.ndarray]] = None,
+    optional_structure_info: tuple,
+    displacement_ids: Sequence | NDArray | None = None,
     zfill_width: int = 3,
-    additional_info: Optional[dict] = None,
+    additional_info: dict | None = None,
 ):
     """Write supercell with displacements to files in each calculator format.
 
@@ -395,6 +403,12 @@ def write_supercells_with_displacements(
         import phonopy.interface.lammps as lammps
 
         lammps.write_supercells_with_displacements(*args, **kwargs)
+
+    elif interface_mode == "qlm":
+        import phonopy.interface.qlm as qlm
+
+        qlm.write_supercells_with_displacements(*args, **kwargs)
+
     else:
         raise RuntimeError("No calculator interface was found.")
 
@@ -418,11 +432,11 @@ def write_magnetic_moments(cell: PhonopyAtoms, sort_by_elements=False):
 
 
 def read_crystal_structure(
-    filename=None,
-    interface_mode=None,
-    chemical_symbols=None,
-    phonopy_yaml_cls: Optional[type[PhonopyYaml]] = None,
-):
+    filename: str | os.PathLike | None = None,
+    interface_mode: str | None = None,
+    chemical_symbols: Sequence[str] | None = None,
+    phonopy_yaml_cls: type[PhonopyYaml] | None = None,
+) -> tuple[PhonopyAtoms, tuple]:
     """Return crystal structure from file in each calculator format.
 
     Parameters
@@ -546,11 +560,18 @@ def read_crystal_structure(
 
         unitcell = read_lammps(cell_filename)
         return unitcell, (cell_filename,)
+
+    elif interface_mode == "qlm":
+        from phonopy.interface.qlm import read_qlm
+
+        unitcell = read_qlm(cell_filename)
+        return unitcell, (cell_filename,)
+
     else:
         raise RuntimeError("No calculator interface was found.")
 
 
-def get_default_cell_filename(interface_mode):
+def get_default_cell_filename(interface_mode: str | None) -> str | None:
     """Return default filename of unit cell structure of each calculator."""
     if interface_mode is None or interface_mode == "vasp":
         return "POSCAR"
@@ -580,13 +601,15 @@ def get_default_cell_filename(interface_mode):
         return "STRU"
     elif interface_mode == "lammps":
         return "unitcell"
+    elif interface_mode == "qlm":
+        return "site"
     elif interface_mode == "pwmat":
         return "atom.config"
     else:
         return None
 
 
-def get_default_supercell_filename(interface_mode):
+def get_default_supercell_filename(interface_mode: str | None) -> str | None:
     """Return default filename of supercell structure of each calculator."""
     if interface_mode == "phonopy_yaml":
         return "phonopy_disp.yaml"
@@ -615,13 +638,15 @@ def get_default_supercell_filename(interface_mode):
         return "sSTRU"
     elif interface_mode == "lammps":
         return "supercell"
+    elif interface_mode == "qlm":
+        return "supercell"
     elif interface_mode == "pwmat":
         return "supercell.config"
     else:
         return None
 
 
-def get_default_displacement_distance(interface_mode):
+def get_default_displacement_distance(interface_mode: str | None) -> float:
     """Return default displacement distance of each calculator."""
     if interface_mode in (
         "wien2k",
@@ -632,6 +657,7 @@ def get_default_displacement_distance(interface_mode):
         "turbomole",
         "fleur",
         "abacus",
+        "qlm",
     ):
         displacement_distance = 0.02
     else:  # default or vasp, crystal, cp2k, pwmat
@@ -639,7 +665,7 @@ def get_default_displacement_distance(interface_mode):
     return displacement_distance
 
 
-def get_default_physical_units(interface_mode=None) -> dict:
+def get_default_physical_units(interface_mode: str | None = None) -> dict:
     """Replace get_calculator_get_physical_units()."""
     warnings.warn(
         (
@@ -652,7 +678,7 @@ def get_default_physical_units(interface_mode=None) -> dict:
     return get_calculator_physical_units(interface_mode=interface_mode)
 
 
-def get_calculator_physical_units(interface_mode=None) -> dict:
+def get_calculator_physical_units(interface_mode: str | None = None) -> dict:
     """Return physical units of each calculator.
 
     Physical units: energy,  distance,  atomic mass, force,        force constants
@@ -671,6 +697,7 @@ def get_calculator_physical_units(interface_mode=None) -> dict:
     fleur         : hartree, au,        AMU,         hartree/au,   hartree/au^2
     abacus        : eV,      au,        AMU,         eV/angstrom,  eV/angstrom.au
     lammps        : eV,      angstrom,  AMU,         eV/angstrom,  eV/angstrom^2
+    qlm           : Ry,      au,        AMU,         Ry/au,        Ry/au^2
     pwmat         : eV,      angstrom,  AMU,         eV/angstrom,  eV/angstrom^2
 
     units['force_constants_unit'] is used in
@@ -837,12 +864,26 @@ def get_calculator_physical_units(interface_mode=None) -> dict:
         units["force_constants_unit"] = "hartree/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "hartree/au"
+    elif interface_mode == "qlm":
+        QlmToTHz = (
+            sqrt(physical_units.Rydberg * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 108.97077
+        units["factor"] = QlmToTHz
+        units["nac_factor"] = 2.0
+        units["distance_to_A"] = physical_units.Bohr
+        units["force_to_eVperA"] = physical_units.Rydberg / physical_units.Bohr
+        units["force_constants_unit"] = "Ry/au^2"
+        units["length_unit"] = "au"
+        units["force_unit"] = "Ry/au"
 
     return units
 
 
 def get_calc_dataset(
-    interface_mode: str,
+    interface_mode: str | None,
     num_atoms: int,
     force_filenames: str,
     verbose: bool = True,
@@ -888,6 +929,8 @@ def get_calc_dataset(
         from phonopy.interface.abacus import parse_set_of_forces
     elif interface_mode == "lammps":
         from phonopy.interface.lammps import parse_set_of_forces
+    elif interface_mode == "qlm":
+        from phonopy.interface.qlm import parse_set_of_forces
     elif interface_mode == "pwmat":
         from phonopy.interface.pwmat import parse_set_of_forces
     else:
@@ -923,7 +966,9 @@ def get_calc_dataset_wien2k(
     return {"forces": force_sets}
 
 
-def get_force_constant_conversion_factor(unit, interface_mode):
+def get_force_constant_conversion_factor(
+    unit: str, interface_mode: str | None
+) -> float:
     """Return unit conversion factor of force constants."""
     _unit = unit.replace("Angstrom", "angstrom")  # for backward compatibility
     interface_default_units = get_calculator_physical_units(interface_mode)
@@ -948,8 +993,8 @@ def get_force_constant_conversion_factor(unit, interface_mode):
 
 
 def _read_phonopy_yaml(
-    filename: str, phonopy_yaml_cls: Optional[type[PhonopyYaml]]
-) -> tuple[PhonopyAtoms, tuple]:
+    filename: str, phonopy_yaml_cls: type[PhonopyYaml] | None
+) -> tuple[PhonopyAtoms | None, tuple]:
     cell_filename = _get_cell_filename(filename, phonopy_yaml_cls)
     if cell_filename is None:
         return None, (None, None)
@@ -970,8 +1015,8 @@ def _read_phonopy_yaml(
 
 
 def _get_cell_filename(
-    filename, phonopy_yaml_cls: Optional[type[PhonopyYaml]]
-) -> Optional[str]:
+    filename: str, phonopy_yaml_cls: type[PhonopyYaml] | None
+) -> str | None:
     cell_filename = None
 
     default_filenames = []